Migrate test_switch_repo to Travis suite
