+++ setup.cfg
@@ -35,3 +35,5 @@ DJANGO_SETTINGS_MODULE=tests.settings
 addopts = --driver Firefox -p no:mozlog -p no:metadata -p no:html
 markers =
     slow: mark a test as slow.
+xfail_strict = true
+sensitive_url = mozilla\.org
