+++ tests/selenium/test_basics.py
@@ -1,7 +1,5 @@
 import datetime
 
-import pytest
-from django.core.management import call_command
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support import expected_conditions as EC
 from selenium.webdriver.support.ui import WebDriverWait
@@ -10,15 +8,8 @@ from treeherder.model.models import (Commit,
                                      Push,
                                      Repository)
 
-pytestmark = pytest.mark.selenium
 
-
-@pytest.fixture
-def initial_data(transactional_db):
-    call_command('load_initial_data')
-
-
-def test_treeherder_main(initial_data, live_server, selenium):
+def test_treeherder_main(live_server, selenium):
     '''
     Tests that the default treeherder page loads ok and we can
     click the repository menu
@@ -31,7 +22,7 @@ def test_treeherder_main(initial_data, live_server, selenium):
     assert selenium.find_element_by_id('repo-dropdown').is_displayed()
 
 
-def test_perfherder_main(initial_data, live_server, selenium):
+def test_perfherder_main(live_server, selenium):
     '''
     This tests that the basic graphs view load and we can click the add tests button
     '''
@@ -52,7 +43,7 @@ def test_perfherder_main(initial_data, live_server, selenium):
     )
 
 
-def test_treeherder_single_commit_titles(initial_data, live_server, selenium):
+def test_treeherder_single_commit_titles(live_server, selenium):
     '''
     This tests that page titles are correct
     '''
