+++ tests/selenium/test_switch_repo.py
@@ -1,12 +1,9 @@
-import pytest
-
-from pages.treeherder import TreeherderPage
+from pages.treeherder import Treeherder
 
 
-@pytest.mark.nondestructive
 def test_switch_repo(base_url, selenium):
     """Switch to new active watched repo"""
-    page = TreeherderPage(selenium, base_url).open()
+    page = Treeherder(selenium, base_url).open()
     assert 'mozilla-inbound' == page.active_watched_repo
     page.select_mozilla_central_repo()
     assert 'mozilla-central' == page.active_watched_repo
